requirements.txt
pandas
requests
pytest
python-dotenv~=0.20